GUI Krypton, InvestManager, справка госслужащего.
* Автозаполнение справки для госслужащих.
* Добавить класс InvestManager, в котором будут парситься и храниться отчеты, инициализироваться все менеджеры и т.п.
* Ошибка в базе - Securities_EconomySectors_Link -> Part должен быть REAL. Переделать скрипт создания БД и запустить заново импорт.
* Переделать GUI на Krypton.

+++ ndp_invest_helper/Models/SecuritiesEconomySectorsLink.cs
@@ -7,7 +7,7 @@ namespace ndp_invest_helper.Models
     {
         public long SecurityId { get; set; }
         public long SectorId { get; set; }
-        public double Part { get; set; }
+        public decimal Part { get; set; }
 
         public virtual EconomySector Sector { get; set; }
         public virtual Security Security { get; set; }
